requirements.txt
Flask
firebase-admin
psycopg2-binary
pytz
gunicorn
